pyproject.toml
[project]
name = "puma-560-pid-rl-research"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.7",
    "numpy<2",
    "roboticstoolbox-python>=1.1.1",
]

[dependency-groups]
dev = [
    "ruff>=0.14.4",
]
